refactor:replace supabase dependency
