+++ packages/database/package.json
@@ -1,9 +1,9 @@
 {
   "name": "@inlang/database",
   "version": "0.0.0",
-  "description": "",
-  "type": "commonjs",
-  "types": "./src/types/definitions.ts",
+  "private": true,
+  "main": "src/index.ts",
+  "types": "src/index.ts",
   "scripts": {
     "start": "docker compose -p inlang-database --file ./supabase/docker/docker-compose.yml up -d && echo \"waiting for database to initialize the schema...\" && sleep 30 && npx prisma migrate dev && npm run apply-functions && npm run apply-triggers && npm run seed && npm run apply-policies",
     "stop": "docker compose -p inlang-database --file ./supabase/docker/docker-compose.yml down",
@@ -17,12 +17,8 @@
     "apply-policies": "ts-node prisma/apply_policies.ts",
     "apply-functions": "ts-node prisma/apply_functions.ts"
   },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
   "devDependencies": {
     "@prisma/client": "^3.2.1",
-    "@supabase/supabase-js": "^1.23.0",
     "@types/jest": "^27.0.1",
     "@types/node": "^16.10.1",
     "jest": "^27.0.6",
@@ -32,5 +28,11 @@
     "ts-jest": "^27.0.4",
     "ts-node": "^10.2.1",
     "typescript": "^4.4.3"
+  },
+  "dependencies": {
+    "@inlang/common": "*",
+    "@inlang/fluent-syntax": "*",
+    "@inlang/adapters": "*",
+    "@supabase/supabase-js": "^1.29.2"
   }
 }
